Move auth related apis from /api to /api/auth
