+++ src/api/auth/login.ts
@@ -1,8 +1,8 @@
 import { Express, Request, Response } from "express";
 import { Db } from "mongodb";
 
-import * as hash from "../utils/hash.js";
-import { isUsersSchema, isUsername, isEmail, isPassword, Username, Email, Password } from "../schema.js";
+import * as hash from "../../utils/hash.js";
+import { isUsersSchema, isUsername, isEmail, isPassword, Username, Email, Password } from "../../schema.js";
 
 // Data required for login: username/email and password.
 interface LoginData {
@@ -35,7 +35,7 @@ function isLoginData(data: unknown): data is LoginData {
  * @param {Db} database - The MongoDB database instance.
  */
 export default (app: Express, database: Db) => {
-    app.post("/api/login", async (req: Request, res: Response) => {
+    app.post("/api/auth/login", async (req: Request, res: Response) => {
         if (isLoginData(req.body)) {
             const { usernameEmail, password } = req.body;
             if (isEmail(usernameEmail)) {
