+++ views/login.ejs
@@ -17,7 +17,7 @@
         <button href="/register">Register</button>
         <script type="module">
             import formSend from "/js/formSend.js";
-            formSend("/api/login", document.getElementById("loginForm"), "/");
+            formSend("/api/auth/login", document.getElementById("loginForm"), "/");
         </script>
     </body>
 </html>
