+++ views/logout.ejs
@@ -13,7 +13,7 @@
         </form>
         <script type="module">
             import formSend from "/js/formSend.js";
-            formSend("/api/logout", document.getElementById("logoutForm"), "/");
+            formSend("/api/auth/logout", document.getElementById("logoutForm"), "/");
         </script>
     </body>
 </html>
