+++ views/register.ejs
@@ -17,7 +17,7 @@
         </form>
         <script type="module">
             import formSend from "/js/formSend.js";
-            formSend("/api/register", document.getElementById("registerForm"), "/login");
+            formSend("/api/auth/register", document.getElementById("registerForm"), "/login");
         </script>
     </body>
 </html>
